+++ Core/Fusion Bundles/Assets/Content/Texture2D/Notifications.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52ac580eaf0c73447b87f6ed952614db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
